move validate_custom_event into new dir
