+++ tests/testing_support/fixtures.py
@@ -894,56 +894,6 @@ def validate_application_exception_message(expected_message):
     return _validate_application_exception_message
 
 
-def _validate_custom_event(recorded_event, required_event):
-    assert len(recorded_event) == 2  # [intrinsic, user attributes]
-
-    intrinsics = recorded_event[0]
-
-    assert intrinsics["type"] == required_event[0]["type"]
-
-    now = time.time()
-    assert isinstance(intrinsics["timestamp"], int)
-    assert intrinsics["timestamp"] <= 1000.0 * now
-    assert intrinsics["timestamp"] >= 1000.0 * required_event[0]["timestamp"]
-
-    assert recorded_event[1].items() == required_event[1].items()
-
-
-def validate_custom_event_in_application_stats_engine(required_event):
-    @function_wrapper
-    def _validate_custom_event_in_application_stats_engine(wrapped, instance, args, kwargs):
-        try:
-            result = wrapped(*args, **kwargs)
-        except:
-            raise
-        else:
-            stats = core_application_stats_engine(None)
-            assert stats.custom_events.num_samples == 1
-
-            custom_event = next(iter(stats.custom_events))
-            _validate_custom_event(custom_event, required_event)
-
-        return result
-
-    return _validate_custom_event_in_application_stats_engine
-
-
-def validate_custom_event_count(count):
-    @function_wrapper
-    def _validate_custom_event_count(wrapped, instance, args, kwargs):
-        try:
-            result = wrapped(*args, **kwargs)
-        except:
-            raise
-        else:
-            stats = core_application_stats_engine(None)
-            assert stats.custom_events.num_samples == count
-
-        return result
-
-    return _validate_custom_event_count
-
-
 def _validate_node_parenting(node, expected_node):
     assert node.exclusive >= 0, "node.exclusive = %s" % node.exclusive
